+++ packages/client/src/publish/MessageFactory.ts
@@ -12,7 +12,7 @@ import {
 import { EncryptionUtil } from '../encryption/EncryptionUtil'
 import { GroupKeyId } from '../encryption/GroupKey'
 import { createMessageRef, createRandomMsgChainId } from './messageChain'
-import { MessageMetadata } from './Publisher'
+import { PublishMetadata } from './Publisher'
 import { keyToArrayIndex } from '@streamr/utils'
 import { GroupKeyQueue } from './GroupKeyQueue'
 import { Mapping } from '../utils/Mapping'
@@ -67,7 +67,7 @@ export class MessageFactory {
     /* eslint-disable padding-line-between-statements */
     async createMessage<T>(
         content: T,
-        metadata: MessageMetadata & { timestamp: number },
+        metadata: PublishMetadata & { timestamp: number },
         explicitPartition?: number
     ): Promise<StreamMessage<T>> {
         const publisherId = await this.authentication.getAddress()
@@ -76,7 +76,7 @@ export class MessageFactory {
             throw new Error(`${publisherId} is not a publisher on stream ${this.streamId}`)
         }
 
-        const partitionCount = (await this.streamRegistry.getStream(this.streamId)).partitions
+        const partitionCount = (await this.streamRegistry.getStream(this.streamId)).getMetadata().partitions
         let partition
         if (explicitPartition !== undefined) {
             if ((explicitPartition < 0 || explicitPartition >= partitionCount)) {