+++ packages/client/test/integration/pre-agreed-encryption-key.test.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata'
 
-import { StreamMessage } from 'streamr-client-protocol'
+import { StreamMessageType } from 'streamr-client-protocol'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { StreamPermission } from '../../src/permission'
 import { nextValue } from '../../src/utils/iterators'
@@ -30,9 +30,9 @@ describe('pre-agreed encryption key', () => {
         await publisher.publish(stream.id, { foo: 'bar' })
         const receivedMessage = await nextValue(sub[Symbol.asyncIterator]())
 
-        expect(receivedMessage?.groupKeyId).toBe(key.id)
+        expect(receivedMessage?.streamMessage.groupKeyId).toBe(key.id)
         const groupKeyRequests = environment.getNetwork().getSentMessages({
-            messageType: StreamMessage.MESSAGE_TYPES.GROUP_KEY_REQUEST
+            messageType: StreamMessageType.GROUP_KEY_REQUEST
         })
         expect(groupKeyRequests).toHaveLength(0)
     })