+++ packages/client/test/integration/sequential-resend-subscribe.test.ts
@@ -1,4 +1,5 @@
 import 'reflect-metadata'
+import { Message } from '../../src/Message'
 
 import { StreamPermission } from '../../src/permission'
 import { Stream } from '../../src/Stream'
@@ -17,9 +18,9 @@ describe('sequential resend subscribe', () => {
     let stream: Stream
 
     let publishTestMessages: ReturnType<typeof getPublishTestStreamMessages>
-    let waitForStorage: (...args: any[]) => Promise<void> = async () => {}
+    let waitForStorage: (msg: Message) => Promise<void> = async () => {}
 
-    let published: any[] = [] // keeps track of stream message data so we can verify they were resent
+    let published: Message[] = [] // keeps track of stream message data so we can verify they were resent
     let environment: FakeEnvironment
 
     beforeAll(async () => {