+++ packages/protocol/src/protocol/message_layer/StreamMessage.ts
@@ -43,7 +43,7 @@ export interface StreamMessageOptions<T> {
     signature: string
 }
 
-export interface ObjectType<T> { 
+export interface ObjectType<T> {
     streamId: string
     streamPartition: number
     timestamp: number
@@ -76,18 +76,9 @@ export type StreamMessageUnencrypted<T> = StreamMessage<T> & {
 export default class StreamMessage<T = unknown> {
     static LATEST_VERSION = LATEST_VERSION
 
-    // TODO can we remove these static field and use the enum object directly?
-    static MESSAGE_TYPES = StreamMessageType
-
-    static VALID_MESSAGE_TYPES = new Set(Object.values(StreamMessage.MESSAGE_TYPES))
-
-    static CONTENT_TYPES = ContentType
-
-    static VALID_CONTENT_TYPES = new Set(Object.values(StreamMessage.CONTENT_TYPES))
-
-    static ENCRYPTION_TYPES = EncryptionType
-
-    static VALID_ENCRYPTIONS = new Set(Object.values(StreamMessage.ENCRYPTION_TYPES))
+    private static VALID_MESSAGE_TYPES = new Set(Object.values(StreamMessageType))
+    private static VALID_CONTENT_TYPES = new Set(Object.values(ContentType))
+    private static VALID_ENCRYPTIONS = new Set(Object.values(EncryptionType))
 
     messageId: MessageID
     prevMsgRef: MessageRef | null
@@ -104,7 +95,7 @@ export default class StreamMessage<T = unknown> {
      * Create a new StreamMessage identical to the passed-in streamMessage.
      */
     clone(): StreamMessage<T> {
-        const content = this.encryptionType === StreamMessage.ENCRYPTION_TYPES.NONE
+        const content = this.encryptionType === EncryptionType.NONE
             ? this.getParsedContent()
             : this.getSerializedContent()
 
@@ -125,9 +116,9 @@ export default class StreamMessage<T = unknown> {
         messageId,
         prevMsgRef = null,
         content,
-        messageType = StreamMessage.MESSAGE_TYPES.MESSAGE,
-        contentType = StreamMessage.CONTENT_TYPES.JSON,
-        encryptionType = StreamMessage.ENCRYPTION_TYPES.NONE,
+        messageType = StreamMessageType.MESSAGE,
+        contentType = ContentType.JSON,
+        encryptionType = EncryptionType.NONE,
         groupKeyId = null,
         newGroupKey = null,
         signature,
@@ -219,12 +210,12 @@ export default class StreamMessage<T = unknown> {
     getParsedContent(): T {
         if (this.parsedContent == null) {
             // Don't try to parse encrypted messages
-            if (this.messageType === StreamMessage.MESSAGE_TYPES.MESSAGE && this.encryptionType !== StreamMessage.ENCRYPTION_TYPES.NONE) {
+            if (this.messageType === StreamMessageType.MESSAGE && this.encryptionType !== EncryptionType.NONE) {
                 // @ts-expect-error need type narrowing for encrypted vs unencrypted
                 return this.serializedContent
             }
 
-            if (this.contentType === StreamMessage.CONTENT_TYPES.JSON) {
+            if (this.contentType === ContentType.JSON) {
                 try {
                     this.parsedContent = JSON.parse(this.serializedContent!)
                 } catch (err: any) {
@@ -371,7 +362,7 @@ export default class StreamMessage<T = unknown> {
             contentType: this.contentType,
             encryptionType: this.encryptionType,
             groupKeyId: this.groupKeyId,
-            content: (this.encryptionType === StreamMessage.ENCRYPTION_TYPES.NONE ? this.getParsedContent() : this.getSerializedContent()),
+            content: (this.encryptionType === EncryptionType.NONE ? this.getParsedContent() : this.getSerializedContent()),
             signature: this.signature,
         }
     }