+++ packages/protocol/test/unit/protocol/control_layer/BroadcastMessageSerializerV2.test.ts
@@ -5,7 +5,8 @@ import {
     BroadcastMessage,
     ControlMessage,
     ContentType,
-    EncryptionType
+    EncryptionType,
+    StreamMessageType
 } from '../../../../src/index'
 import { toEthereumAddress } from '@streamr/utils'
 import { SIGNATURE_TYPE_ETH } from '../../../../src/protocol/message_layer/StreamMessageSerializerV32'
@@ -13,7 +14,7 @@ import { SIGNATURE_TYPE_ETH } from '../../../../src/protocol/message_layer/Strea
 const PUBLISHER_ID = toEthereumAddress('0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
 
 const streamMessage = StreamMessage.deserialize([32, ['streamId', 0, 1529549961116, 0, PUBLISHER_ID, 'msg-chain-id'],
-    [1529549961000, 0], StreamMessage.MESSAGE_TYPES.MESSAGE,
+    [1529549961000, 0], StreamMessageType.MESSAGE,
     ContentType.JSON, EncryptionType.NONE, null, '{"valid": "json"}', null, SIGNATURE_TYPE_ETH, 'signature'])
 
 const VERSION = 2