+++ packages/client/CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Methods related to publishing and subscribing operate on new interfaces `Message` and `MessageMetadata` instead of `StreamMessage`
+  - in `client.subscribe` and `client.resend` the data type of 2nd parameter of `onMessage` callback is `MessageMetadata` instead of `StreamMessage`
+  - in `client.subscribe`, `client.resend` and `client.resendSubscribe` the async iterator type is `Message` instead of `StreamMessage`
+  - in `client.publish` and `stream.publish` the return type is `Message` instead of `StreamMessage`
+  - in `client.waitForStorage` parameter type is `Message` instead of `StreamMessage`
 - Encryption keys are delivered in-stream, not in a separate key exchange stream
   - new optional config options `decryption.keyRequestTimeout` and `decryption.maxKeyRequestsPerSecond`
   - notice that key exchange is not backwards compatible with v6 clients
@@ -25,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enforce concurrency limit for smart contract calls (per contract, configurable with `maxConcurrentContractCalls` config option)
 - Enforce presence of message signatures
   - all non-signed messages received by client are simply ignored
+- Stream metadata now accessed through `stream.getMetadata()`
+  - e.g. usages of `stream.partitions` has changed to `stream.getMetadata().partitions`
 - Method `stream.update()` parameter `props` is no longer optional
 - Rename method `getStorageNodesOf()` to `getStorageNodes()`
 - Rename method `getStoredStreamsOf()` to `getStoredStreams()`
@@ -39,8 +46,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change interface of `MessageStream` from `AsyncGenerator` to `AsyncIterable`
 - Change return type of `getStreamPublishers`, `getStreamSubscribers` and `searchStreams` from `AsyncGenerator` to `AsyncIterable`
 - Config property `auth` must be non-empty (if given)
+- Result set of `getStoredStreams` is no longer capped to 1000 streams
 
 ### Deprecated
+
+- Deprecate `client.getNode` method and interface `NetworkNodeStub`
 
 ### Removed
 
@@ -52,6 +62,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - use `destroy()` instead
 - Remove method `unsubscribeAll()`
   - use `unsubscribe()` without arguments to same effect
+- Remove method `stream.toObject()` and interface `StreamProperties`
+  - use `stream.getMetadata()` to get metadata (doesn't contain stream id)
+  - use interface `StreamMetadata` instead
 - Remove properties `subscription.onMessage`, `onStart`, and `onError`
   - use `subscription.on('error', cb)` to add an error listener
 - Remove configuration option `groupKeys`